refactor test orginization/layout
This refactoring makes use to the new "relaxed"
test layout and the integration and end-to-end markers so that pyest
by default only runs unit tests.

A lot more work is needed to fully implement end-to-end and
integration testing with external resources
(REST APIs, SSH servers, Databases, etc), but this is a good start.
