+++ projects/nautobot/uoft_nautobot/tests/end_to_end.py
@@ -1,7 +1,8 @@
 from ..librenms import Settings
 import pytest
 
-@pytest.mark.skip(reason="WIP")
+@pytest.mark.end_to_end
 class LibreNMS:
     def happy_path(self, mocker):
         s = Settings.from_cache()
+        print(s)
