+++ projects/switchconfig/uoft_switchconfig/tests/end_to_end.py
@@ -1,0 +1,4 @@
+import pytest
+
+pytestmark = pytest.mark.end_to_end
+
