+++ tests/test_ensure_python.py
@@ -1,2 +1,3 @@
 "run tasks/ensure-python.py in a bunch of different docker containers, and make assertions about the results"
 
+#TODO: add a test that runs the script in a docker container, and makes assertions about the results
